cmd/internal/gc: make class uint8 throughout
This prevents conversion noise.

No functional changes. Passes toolstash -cmp.

Reviewed-on: https://go-review.googlesource.com/8164

+++ src/cmd/internal/gc/export.go
@@ -57,7 +57,7 @@ func exportedsym(sym *Sym) bool {
 	return sym.Pkg == localpkg && exportname(sym.Name)
 }
 
-func autoexport(n *Node, ctxt int) {
+func autoexport(n *Node, ctxt uint8) {
 	if n == nil || n.Sym == nil {
 		return
 	}
